Putting all test helper functions in conftest.py. So far I've moved the fixtures that were already in test_deadwood_litter_rasterio.py into conftest and everything seems to still be working.
